+++ src/WeightTemplates/BardTemplate.java
@@ -5,44 +5,44 @@ import java.util.ArrayList;
 public class BardTemplate extends WeightTemplate{
 
 	@Override
-	public ArrayList<Integer> getAggressiveTemplate() {
-		ArrayList<Integer> template = new ArrayList<Integer>();
-		template.add(70);
-		template.add(15);
-		template.add(15);
+	public ArrayList<Double> getAggressiveTemplate() {
+		ArrayList<Double> template = new ArrayList<Double>();
+		template.add(70.0);
+		template.add(15.0);
+		template.add(15.0);
 		
-		template.add(30);
-		template.add(15);
-		template.add(15);
-		template.add(40);
+		template.add(30.0);
+		template.add(15.0);
+		template.add(15.0);
+		template.add(40.0);
 		return template;
 	}
 
 	@Override
-	public ArrayList<Integer> getDefensiveTemplate() {
-		ArrayList<Integer> template = new ArrayList<Integer>();
-		template.add(10);
-		template.add(45);
-		template.add(45);
+	public ArrayList<Double> getDefensiveTemplate() {
+		ArrayList<Double> template = new ArrayList<Double>();
+		template.add(10.0);
+		template.add(45.0);
+		template.add(45.0);
 		
-		template.add(15);
-		template.add(15);
-		template.add(60);
-		template.add(10);
+		template.add(15.0);
+		template.add(15.0);
+		template.add(60.0);
+		template.add(10.0);
 		return template;
 	}
 
 	@Override
-	public ArrayList<Integer> getCCTemplate() {
-		ArrayList<Integer> template = new ArrayList<Integer>();
-		template.add(33);
-		template.add(33);
-		template.add(34);
+	public ArrayList<Double> getCCTemplate() {
+		ArrayList<Double> template = new ArrayList<Double>();
+		template.add(33.0);
+		template.add(33.0);
+		template.add(34.0);
 		
-		template.add(40);
-		template.add(10);
-		template.add(10);
-		template.add(40);
+		template.add(40.0);
+		template.add(10.0);
+		template.add(10.0);
+		template.add(40.0);
 		return template;
 	}
