+++ src/WeightTemplates/WarlockTemplate.java
@@ -5,44 +5,44 @@ import java.util.ArrayList;
 public class WarlockTemplate extends WeightTemplate{
 
 	@Override
-	public ArrayList<Integer> getAggressiveTemplate() {
-		ArrayList<Integer> template = new ArrayList<Integer>();
-		template.add(40);
-		template.add(40);
-		template.add(20);
+	public ArrayList<Double> getAggressiveTemplate() {
+		ArrayList<Double> template = new ArrayList<Double>();
+		template.add(40.0);
+		template.add(40.0);
+		template.add(20.0);
 		
-		template.add(30);
-		template.add(30);
-		template.add(10);
-		template.add(30);
+		template.add(30.0);
+		template.add(30.0);
+		template.add(10.0);
+		template.add(30.0);
 		return template;
 	}
 
 	@Override
-	public ArrayList<Integer> getDefensiveTemplate() {
-		ArrayList<Integer> template = new ArrayList<Integer>();
-		template.add(30);
-		template.add(30);
-		template.add(40);
+	public ArrayList<Double> getDefensiveTemplate() {
+		ArrayList<Double> template = new ArrayList<Double>();
+		template.add(30.0);
+		template.add(30.0);
+		template.add(40.0);
 		
-		template.add(40);
-		template.add(5);
-		template.add(35);
-		template.add(20);
+		template.add(40.0);
+		template.add(5.0);
+		template.add(35.0);
+		template.add(20.0);
 		return template;
 	}
 
 	@Override
-	public ArrayList<Integer> getCCTemplate() {
-		ArrayList<Integer> template = new ArrayList<Integer>();
-		template.add(10);
-		template.add(10);
-		template.add(80);
+	public ArrayList<Double> getCCTemplate() {
+		ArrayList<Double> template = new ArrayList<Double>();
+		template.add(10.0);
+		template.add(10.0);
+		template.add(80.0);
 		
-		template.add(5);
-		template.add(5);
-		template.add(60);
-		template.add(30);
+		template.add(5.0);
+		template.add(5.0);
+		template.add(60.0);
+		template.add(30.0);
 		return template;
 	}
