+++ src/WeightTemplates/WeightTemplate.java
@@ -4,22 +4,22 @@ import java.util.ArrayList;
 
 public abstract class WeightTemplate {
 	
-	public abstract ArrayList<Integer> getAggressiveTemplate();
+	public abstract ArrayList<Double> getAggressiveTemplate();
 	
-	public abstract ArrayList<Integer> getDefensiveTemplate();
+	public abstract ArrayList<Double> getDefensiveTemplate();
 	
-	public ArrayList<Integer> getBalancedTemplate() {
-		ArrayList<Integer> template = new ArrayList<Integer>();
-		template.add(33);
-		template.add(33);
-		template.add(34);
-		template.add(25);
-		template.add(25);
-		template.add(25);
-		template.add(25);
+	public ArrayList<Double> getBalancedTemplate() {
+		ArrayList<Double> template = new ArrayList<Double>();
+		template.add(33.0);
+		template.add(33.0);
+		template.add(34.0);
+		template.add(25.0);
+		template.add(25.0);
+		template.add(25.0);
+		template.add(25.0);
 		return template;
 	}
 	
-	public abstract ArrayList<Integer> getCCTemplate();
+	public abstract ArrayList<Double> getCCTemplate();
 	
 }
